Split cook's assistant and witch's potion item retrieve leafs into their respective leafs. Fixed Witchs Potion lol
